# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Api to create a language ll1",
        "title": "GO-Api API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/scanners/priv/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a Private scanner from a Yalex",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Create a Private scanner from a Yalex",
                "parameters": [
                    {
                        "description": "Yalex object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.YalexPublic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/priv/simulate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Simulate a scanner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Simulate a scanner",
                "parameters": [
                    {
                        "description": "Scanner object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.SimScan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/public/create": {
            "post": {
                "description": "Create a Public scanner from a Yalex",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Create a Public scanner from a Yalex",
                "parameters": [
                    {
                        "description": "Yalex object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.YalexPublic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/public/simulate": {
            "post": {
                "description": "Simulate a scanner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Simulate a scanner",
                "parameters": [
                    {
                        "description": "Scanner object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.SimScan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
...
        "/image/{fileID}": {
            "get": {
                "description": "Get an image from GridFS by file ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Get an image from GridFS by file ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "afd.ResponseSim": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "simulatedParts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/afd.SimulatedPart"
                    }
                },
                "stringSummary": {
                    "type": "string"
                }
            }
        },
        "afd.SimulatedPart": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "final": {
                    "type": "integer"
                },
                "init": {
                    "type": "integer"
                },
                "runes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "scanners.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "scanners.ScannerRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "scannerName": {
                    "type": "string"
                }
            }
        },
        "scanners.ScannerResponse": {
            "type": "object",
            "properties": {
                "filesId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Scanner created successfully"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "scanners.SimScan": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "simPart": {
                    "$ref": "#/definitions/afd.ResponseSim"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "scanners.YalexPublic": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}